(PC-28265)[PRO] refactor: make AccessibilitySummarySection more generic
