--- pro/src/components/AccessibilitySummarySection/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AccessibilitySummarySection'
