+++ pro/src/screens/IndividualOffer/SummaryScreen/__specs__/SummaryScreen.spec.tsx
@@ -169,7 +169,7 @@ describe('Summary', () => {
     expect(screen.getByText('Type d’offre')).toBeInTheDocument()
     expect(screen.getByText('Informations artistiques')).toBeInTheDocument()
     expect(screen.getByText('Informations pratiques')).toBeInTheDocument()
-    expect(screen.getByText('Accessibilité')).toBeInTheDocument()
+    expect(screen.getByText('Modalités d’accessibilité')).toBeInTheDocument()
     expect(
       screen.getByText('Notifications des réservations')
     ).toBeInTheDocument()
